@@ -2,12 +2,12 @@ package proxy
 
 import (
 	"crypto/tls"
+	"fmt"
 	"net"
 	"time"
 
 	"github.com/AdguardTeam/dnsproxy/proxyutil"
 	"github.com/AdguardTeam/golibs/log"
-	"github.com/joomcode/errorx"
 	"github.com/miekg/dns"
 )
 
@@ -16,8 +16,9 @@ func (p *Proxy) createTCPListeners() error {
 		log.Printf("Creating a TCP server socket")
 		tcpListen, err := net.ListenTCP("tcp", a)
 		if err != nil {
-			return errorx.Decorate(err, "couldn't listen to TCP socket")
+			return fmt.Errorf("starting listening on tcp socket: %w", err)
 		}
+
 		p.tcpListen = append(p.tcpListen, tcpListen)
 		log.Printf("Listening to tcp://%s", tcpListen.Addr())
 	}
@@ -29,7 +30,7 @@ func (p *Proxy) createTLSListeners() error {
 		log.Printf("Creating a TLS server socket")
 		tcpListen, err := net.ListenTCP("tcp", a)
 		if err != nil {
-			return errorx.Decorate(err, "could not start TLS listener")
+			return fmt.Errorf("starting tls listener: %w", err)
 		}
 
 		l := tls.NewListener(tcpListen, p.TLSConfig)
@@ -68,8 +69,15 @@ func (p *Proxy) tcpPacketLoop(l net.Listener, proto Proto, requestGoroutinesSema
 // handleTCPConnection starts a loop that handles an incoming TCP connection
 // proto is either "tcp" or "tls"
 func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
-	log.Tracef("Start handling the new %s connection %s", proto, conn.RemoteAddr())
-	defer conn.Close()
+	defer log.OnPanic("proxy.handleTCPConnection")
+
+	log.Tracef("handling tcp: started handling %s request from %s", proto, conn.RemoteAddr())
+	defer func() {
+		err := conn.Close()
+		if err != nil {
+			log.Error("handling tcp: closing conn: %s", err)
+		}
+	}()
 
 	for {
 		p.RLock()
@@ -78,16 +86,24 @@ func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 		}
 		p.RUnlock()
 
-		conn.SetDeadline(time.Now().Add(defaultTimeout)) //nolint
+		err := conn.SetDeadline(time.Now().Add(defaultTimeout))
+		if err != nil {
+			// Consider deadline errors non-critical.
+			log.Error("handling tcp: setting deadline: %s", err)
+		}
+
 		packet, err := proxyutil.ReadPrefixed(conn)
 		if err != nil {
+			log.Error("handling tcp: reading: %s", err)
+
 			return
 		}
 
 		req := &dns.Msg{}
 		err = req.Unpack(packet)
 		if err != nil {
-			log.Info("error handling TCP packet: %s", err)
+			log.Error("handling tcp: unpacking: %s", err)
+
 			return
 		}
 
@@ -97,7 +113,7 @@ func (p *Proxy) handleTCPConnection(conn net.Conn, proto Proto) {
 
 		err = p.handleDNSRequest(d)
 		if err != nil {
-			log.Tracef("error handling DNS (%s) request: %s", d.Proto, err)
+			log.Error("handling tcp: handling %s request: %s", d.Proto, err)
 		}
 	}
 }
@@ -114,16 +130,15 @@ func (p *Proxy) respondTCP(d *DNSContext) error {
 
 	bytes, err := resp.Pack()
 	if err != nil {
-		return errorx.Decorate(err, "couldn't convert message into wire format: %s", resp.String())
+		return fmt.Errorf("packing message: %w", err)
 	}
 
 	err = proxyutil.WritePrefixed(bytes, conn)
-
 	if proxyutil.IsConnClosed(err) {
 		return err
 	}
 	if err != nil {
-		return errorx.Decorate(err, "conn.Write() returned error")
+		return fmt.Errorf("conn.Write(): %w", err)
 	}
 
 	return nil
