@@ -3,6 +3,7 @@ package proxyutil
 import (
 	"encoding/binary"
 	"errors"
+	"fmt"
 	"io"
 	"net"
 
@@ -32,13 +33,15 @@ func DNSSize(isUDP bool, r *dns.Msg) int {
 	return int(size)
 }
 
-// ReadPrefixed -- reads a DNS message with a 2-byte prefix containing message length
+// ReadPrefixed reads a DNS message with a 2-byte prefix containing message
+// length from conn.
 func ReadPrefixed(conn net.Conn) ([]byte, error) {
 	l := make([]byte, 2)
 	_, err := conn.Read(l)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("reading len: %w", err)
 	}
+
 	packetLen := binary.BigEndian.Uint16(l)
 	if packetLen > dns.MaxMsgSize {
 		return nil, ErrTooLarge
@@ -47,8 +50,9 @@ func ReadPrefixed(conn net.Conn) ([]byte, error) {
 	buf := make([]byte, packetLen)
 	_, err = io.ReadFull(conn, buf)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("reading msg: %w", err)
 	}
+
 	return buf, nil
 }
 
