@@ -13,7 +13,6 @@ import (
 	"github.com/AdguardTeam/golibs/log"
 	"github.com/ameshkov/dnscrypt/v2"
 	"github.com/ameshkov/dnsstamps"
-	"github.com/joomcode/errorx"
 	"github.com/miekg/dns"
 )
 
@@ -79,8 +78,9 @@ func AddressToUpstream(address string, options *Options) (Upstream, error) {
 	if strings.Contains(address, "://") {
 		upstreamURL, err := url.Parse(address)
 		if err != nil {
-			return nil, errorx.Decorate(err, "failed to parse %s", address)
+			return nil, fmt.Errorf("failed to parse %s: %w", address, err)
 		}
+
 		return urlToUpstream(upstreamURL, options)
 	}
 
@@ -129,7 +129,7 @@ func urlToUpstream(upstreamURL *url.URL, opts *Options) (Upstream, error) {
 
 		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
-			return nil, errorx.Decorate(err, "couldn't create quic bootstrapper")
+			return nil, fmt.Errorf("creating quic bootstrapper: %w", err)
 		}
 
 		return &dnsOverQUIC{boot: b}, nil
@@ -141,7 +141,7 @@ func urlToUpstream(upstreamURL *url.URL, opts *Options) (Upstream, error) {
 
 		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
-			return nil, errorx.Decorate(err, "couldn't create tls bootstrapper")
+			return nil, fmt.Errorf("creating tls bootstrapper: %w", err)
 		}
 
 		return &dnsOverTLS{boot: b}, nil
@@ -153,7 +153,7 @@ func urlToUpstream(upstreamURL *url.URL, opts *Options) (Upstream, error) {
 
 		b, err := urlToBoot(upstreamURL, opts)
 		if err != nil {
-			return nil, errorx.Decorate(err, "couldn't create tls bootstrapper")
+			return nil, fmt.Errorf("creating https bootstrapper: %w", err)
 		}
 
 		return &dnsOverHTTPS{boot: b}, nil
@@ -168,7 +168,7 @@ func urlToUpstream(upstreamURL *url.URL, opts *Options) (Upstream, error) {
 func stampToUpstream(upsURL *url.URL, opts *Options) (Upstream, error) {
 	stamp, err := dnsstamps.NewServerStampFromString(upsURL.String())
 	if err != nil {
-		return nil, errorx.Decorate(err, "failed to parse %s", upsURL)
+		return nil, fmt.Errorf("failed to parse %s: %w", upsURL, err)
 	}
 
 	if stamp.ServerAddrStr != "" {
